fix: 상품 삭제

+++ src/main/java/com/age/b2b/service/ProductService.java
@@ -4,6 +4,8 @@ import com.age.b2b.domain.Product;
 import com.age.b2b.domain.common.ProductStatus;
 import com.age.b2b.dto.ProductRequestDto;
 import com.age.b2b.dto.ProductResponseDto;
+import com.age.b2b.repository.CartItemRepository;
+import com.age.b2b.repository.OrderItemRepository;
 import com.age.b2b.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -25,6 +27,8 @@ import java.util.stream.Collectors;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final CartItemRepository cartItemRepository;
+    private final OrderItemRepository orderItemRepository;
 
     @Transactional(readOnly = true)
     public String getProductNameByCode(String code) {
@@ -56,7 +60,7 @@ public class ProductService {
 
     // 1. 상품 등록
     public Long saveProduct(ProductRequestDto dto) {
-        // ★ 1. 상품코드 자동 생성 로직 (P + 년월일시분초 + 3자리난수)
+        // 1. 상품코드 자동 생성 로직 (P + 년월일시분초 + 3자리난수)
         // 예: P20231217103000123
         String generatedCode = generateProductCode();
 
@@ -76,13 +80,13 @@ public class ProductService {
         product.setDescription(dto.getDescription());
         product.setExpiryDate(dto.getExpiryDate());
 
-        // ★ 2. DTO에서 변환된 상태값 사용
+        // 2. DTO에서 변환된 상태값 사용
         product.setStatus(dto.getStatus() != null ? dto.getStatus() : ProductStatus.ON_SALE);
 
         return productRepository.save(product).getId();
     }
 
-    // 2. 상품 수정 (Dirty Checking)
+    // 2. 상품 수정
     public void updateProduct(Long id, ProductRequestDto dto) {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("상품이 없습니다."));
@@ -103,6 +107,15 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        // (1) 주문 내역 확인: 이미 팔린 상품은 삭제 불가
+        if (orderItemRepository.existsByProductId(id)) {
+            throw new IllegalStateException("이미 주문 이력이 있는 상품이 포함되어 있습니다.\n상품 수정에서 상품상태를 변경해주세요.");
+        }
+
+        // (2) 장바구니 내역 삭제
+        cartItemRepository.deleteByProductId(id);
+
+        // (3) 상품 삭제
         productRepository.deleteById(id);
     }
